flask-backend/requirements.txt
Flask==2.1.2
Werkzeug==2.0.3
kafka-python==2.0.2